Add files via upload
I figured out that we needed to change the socket/ TCP connection in the client file to an HTTP connection, and I also figured out how to send a post request. I'm not sure how to read HTTP responses thought so that part of the code is not finished. I added a way to make an HTML table to the code to but I don't think it is correct.